refactor passing tests
