+++ conan/internal/api/list/query_parse.py
@@ -1,14 +1,59 @@
+from collections import OrderedDict
 
-def is_operator(el):
+
+def filter_package_configs(pkg_configurations, query):
+    postfix = _infix_to_postfix(query) if query else []
+    result = OrderedDict()
+    for pref, data in pkg_configurations.items():
+        if _evaluate_postfix_with_info(postfix, data):
+            result[pref] = data
+    return result
+
+
+def _evaluate_postfix_with_info(postfix, binary_info):
+
+    # Evaluate conaninfo with the expression
+
+    def evaluate_info(expression):
+        """Receives an expression like compiler.version="12"
+        Uses conan_vars_info in the closure to evaluate it"""
+        name, value = expression.split("=", 1)
+        value = value.replace("\"", "")
+        return _evaluate(name, value, binary_info)
+
+    return _evaluate_postfix(postfix, evaluate_info)
+
+
+def _evaluate(prop_name, prop_value, binary_info):
+    """
+    Evaluates a single prop_name, prop_value like "os", "Windows" against
+    conan_vars_info.serialize_min()
+    """
+
+    def compatible_prop(setting_value, _prop_value):
+        return (_prop_value == setting_value) or (_prop_value == "None" and setting_value is None)
+
+    # TODO: Necessary to generalize this query evaluation to include all possible fields
+    info_settings = binary_info.get("settings", {})
+    info_options = binary_info.get("options", {})
+
+    if not prop_name.startswith("options."):
+        return compatible_prop(info_settings.get(prop_name), prop_value)
+    else:
+        prop_name = prop_name[len("options."):]
+        return compatible_prop(info_options.get(prop_name), prop_value)
+
+
+def _is_operator(el):
     return el in ["|", "&"]
 
 
 def _parse_expression(subexp):
-    '''Expressions like:
+    """Expressions like:
      compiler.version=12
      compiler="Visual Studio"
      arch="x86"
-     Could be replaced with another one to parse different queries'''
+     Could be replaced with another one to parse different queries """
     ret = ""
     quoted = False
     for char in subexp:
@@ -19,7 +64,7 @@ def _parse_expression(subexp):
 
         if quoted:
             ret += char
-        elif char == " " or is_operator(char) or char in [")", "("]:
+        elif char == " " or _is_operator(char) or char in [")", "("]:
             break
         else:
             ret += char
@@ -30,7 +75,7 @@ def _parse_expression(subexp):
     return ret
 
 
-def evaluate_postfix(postfix, evaluator):
+def _evaluate_postfix(postfix, evaluator):
     """
     Evaluates a postfix expression and returns a boolean
     @param postfix:  Postfix expression as a list
@@ -43,7 +88,7 @@ def evaluate_postfix(postfix, evaluator):
 
     stack = []
     for el in postfix:
-        if not is_operator(el):
+        if not _is_operator(el):
             stack.append(el)
         else:
             o1 = stack.pop()
@@ -66,7 +111,7 @@ def evaluate_postfix(postfix, evaluator):
         return stack[0]
 
 
-def infix_to_postfix(exp):
+def _infix_to_postfix(exp):
     """
     Translates an infix expression to postfix using an standard algorithm
     with little hacks for parse complex expressions like "compiler.version=4"
@@ -98,11 +143,11 @@ def infix_to_postfix(exp):
                     output.append(popped)
             if popped != "(":
                 raise Exception("Bad expression, not balanced parenthesis")
-        elif is_operator(char):
+        elif _is_operator(char):
             # Same operations has the same priority
             # replace this lines if the operators need to have
             # some priority
-            if stack and is_operator(stack[:-1]):
+            if stack and _is_operator(stack[:-1]):
                 popped = stack.pop()
                 output.append(popped)
             stack.append(char)
